Relocate Types component
